requirements.txt
streamlit
pandas
openpyxl
python-decouple
psycopg2-binary

